% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elevation}
\alias{elevation}
\title{Elevation raster data}
\format{
A RasterLayer object
}
\source{
\url{http://srtm.csi.cgiar.org/SRT-ZIP/SRTM_V41/SRTM_Data_GeoTiff/srtm_14_05.zip}
}
\usage{
elevation
}
\description{
Elevation raster data from SRTM of the Zion National Park area
}
\keyword{datasets}
