% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{landsat}
\alias{landsat}
\title{Dataset landsat}
\format{A multilayer geotiff file}
\source{
\url{http://landsat-pds.s3.amazonaws.com/L8/038/034/LC80380342015230LGN00/}
}
\description{
This is a dataset containing the four bands (2, 3, 4, 5) of the Landsat 8 image for
the area of Zion National Park
}
\section{landsat.tif}{

}

