% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{landsat_b2}
\alias{landsat_b2}
\title{Dataset landsat_b2}
\format{A geotiff file}
\source{
\url{http://landsat-pds.s3.amazonaws.com/L8/038/034/LC80380342015230LGN00/}
}
\description{
This is a dataset containing the second band of the Landsat 8 image for
the area of Zion National Park
}
