% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcd}
\alias{nlcd}
\title{Land cover raster data}
\format{A RasterLayer object
\itemize{
        \item{1} {Water}
        \item{2} {Developed}
        \item{3} {Barren}
        \item{4} {Forest}
        \item{5} {Shrubland}
        \item{6} {Herbaceous}
        \item{7} {Cultivated}
        \item{8} {Wetlands}
}}
\source{
\url{https://www.mrlc.gov/nlcd2011.php}
}
\usage{
nlcd
}
\description{
This is a dataset containing the National Land Cover Database 2011 product for
the Zion National Park area simplified to eight land cover categories
}
\keyword{datasets}
