% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{nlcd2011.tif}
\alias{nlcd2011.tif}
\title{Dataset nlcd2011}
\format{
A geotiff file
}
\source{
\url{https://www.mrlc.gov/nlcd2011.php}
}
\description{
This is a dataset containing the National Land Cover Database 2011 product for
the Zion National Park area
}
\examples{

system.file("raster/nlcd2011.tif", package = "spDataLarge")

}
