% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_elev}
\alias{nz_elev}
\title{New Zeleand elevation raster data}
\format{A RasterLayer object}
\source{
\url{https://aws.amazon.com/public-datasets/terrain/}
}
\usage{
nz_elev
}
\description{
Elevation raster data of the New Zeleand area from the Mapzen Terrain Service
}
\seealso{
See the elevatr package: https://cran.r-project.org/web/packages/elevatr
}
\keyword{datasets}
