% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{nz_elev.tif}
\alias{nz_elev.tif}
\title{New Zeleand elevation raster data}
\format{
A geotiff file
}
\source{
\url{https://aws.amazon.com/public-datasets/terrain/}
}
\description{
Elevation raster data of the New Zealand area from the Mapzen Terrain Service.
For teaching purposes only
}
\examples{

system.file("raster/nz_elev.tif", package = "spDataLarge")

}
\seealso{
See the elevatr package: https://cran.r-project.org/web/packages/elevatr
}
