% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_download.R
\name{sp_download}
\alias{sp_download}
\title{sp_download}
\usage{
sp_download(filenames, version_id, dest_folder = NULL)
}
\arguments{
\item{filenames}{names of files}

\item{dest_folder}{optional destination folder}
}
\description{
Retrieves external files from the spDataLarge package
}
\examples{
\dontrun{
sp_download(c("zion.gpkg", "landsat_b3.tif"))
}

}
