% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{srtm}
\alias{srtm}
\title{Dataset srtm}
\format{
A geotiff file
}
\source{
\url{http://srtm.csi.cgiar.org/SRT-ZIP/SRTM_V41/SRTM_Data_GeoTiff/srtm_14_05.zip}
}
\description{
This is a dataset containing the elevation raster data from SRTM of
the Zion National Park area
}
