% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landslides.R
\name{ta.tif}
\alias{ta.tif}
\title{Terrain attributes of the landslide dataset from Southern Ecuador}
\format{
A geotiff file with the five terrain attribute layers: slope, plan
  curvature, profile curvature, elevation and catchment area.
}
\source{
Landslide dataset of the RSAGA package: \code{data("landslides",
  package = "RSAGA")}.

\strong{DEM:}

  Ungerechts, L. (2010): DEM 10m (triangulated from aerial photo - b/w).
  Available online:

  `http://www.tropicalmountainforest.org/data_pre.do?citid=901`

  Jordan, E., Ungerechts, L., Caceres, B. Penafiel, A. and Francou, B.
  (2005): Estimation by photogrammetry of the glacier recession on the
  Cotopaxi Volcano (Ecuador) between 1956 and 1997. *Hydrological
  Sciences* 50, 949-961.
}
\description{
Data used in the "Statistical learning for geographic data" chapter in Geocomputation with R.
See \url{http://geocompr.robinlovelace.net/spatial-cv.html} for details.
}
\examples{
terra::rast(system.file("raster/ta.tif", package = "spDataLarge"))
}
\seealso{
\code{?lsl}
}
