% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{zion}
\alias{zion}
\title{Dataset zion}
\format{A geopackage file}
\source{
\url{https://irma.nps.gov/DataStore/DownloadFile/580617}
}
\description{
This is a dataset containing the borders of Zion National Park
}
\section{zion.gpkg}{

}

