% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{zion_points.gpkg}
\alias{zion_points.gpkg}
\title{Point vector data}
\format{
A geopackage file
}
\source{
\url{https://irma.nps.gov/DataStore/DownloadFile/580617}
}
\description{
Dataset containing 30 randomly located points in the Zion National Park
}
\examples{

system.file("vector/zion_points.gpkg", package = "spDataLarge")

}
