% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zion_points}
\alias{zion_points}
\title{Point vector data}
\format{
A sf object
}
\usage{
zion_points
}
\description{
Dataset containing 30 randomly located points in the Zion National Park
}
\keyword{datasets}
