% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristol_ways.R
\docType{data}
\name{bristol_ways}
\alias{bristol_ways}
\alias{bristol_od}
\alias{bristol_region}
\alias{bristol_cents}
\alias{bristol_ttwa}
\alias{bristol_zones}
\alias{bristol_stations}
\title{Datasets providing a snapshot of Bristol's transport system}
\format{
sf data frame objects
}
\source{
\url{http://wicid.ukdataservice.ac.uk/} and other open access sources
}
\usage{
bristol_ways
}
\description{
Data used in the transport chapter in Geocomputation with R.
See \url{http://geocompr.robinlovelace.net/transport.html} for details.
}
\examples{
\dontrun{
library(sf)
bristol_ways
bristol_od
bristol_region
bristol_ttwa
bristol_zones
bristol_stations
}
}
\keyword{datasets}
